Lab10/hh_back/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
drf-spectacular==0.28.0